[Silver II] Title: 과자 나눠주기, Time: 548 ms, Memory: 162452 KB -BaekjoonHub

+++ 백준/Silver/16401. 과자 나눠주기/과자 나눠주기.java
@@ -4,7 +4,7 @@ import java.util.*;
 public class Main {
     static int m;
     static int n;
-    static long[] snack;
+    static int[] snack;
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
@@ -13,23 +13,23 @@ public class Main {
         m = Integer.parseInt(stk.nextToken());
         n = Integer.parseInt(stk.nextToken());
 
-        long start = 1;
-        long end = 1;
+        int start = 1;
+        int end = 1;
 
-        snack = new long[n];
+        snack = new int[n];
         stk = new StringTokenizer(br.readLine());
         for(int i=0;i<n;i++){
-            snack[i] = Long.parseLong(stk.nextToken());
+            snack[i] = Integer.parseInt(stk.nextToken());
             end = Math.max(end, snack[i]);
         }
-        long result = 0;
+        int result = 0;
         while(start <= end){
-            long mid = start + (end - start) / 2L;
+            int mid = start + (end - start) / 2;
             if(check(mid)){
-                start = mid + 1L;
+                start = mid + 1;
                 result = mid;
             }else{
-                end = mid - 1L;
+                end = mid - 1;
             }
         }
         bw.write(String.valueOf(result));
@@ -37,8 +37,8 @@ public class Main {
         bw.close();
         br.close();
     }
-    static boolean check(long mid){
-        long sum = 0;
+    static boolean check(int mid){
+        int sum = 0;
         for(int i=0;i<n;i++){
             sum += snack[i] / mid;
             if(sum >= m) break;
